# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tasks": {
      "post": {
        "operationId": "TaskController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      },
      "get": {
        "operationId": "TaskController_list",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeDeleted",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTasksResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{id}/completions": {
      "post": {
        "operationId": "TaskController_logCompletion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogCompletionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCompletionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      },
      "get": {
        "operationId": "TaskController_listCompletions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCompletionsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{id}": {
      "get": {
        "operationId": "TaskController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      },
      "patch": {
        "operationId": "TaskController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      },
      "delete": {
        "operationId": "TaskController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Habit Tracker API",
    "description": "API for habit tracking, community, challenges and gamification",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TaskReminderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "time": {
            "type": "string"
          },
          "locationName": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "isEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "type",
          "isEnabled"
        ]
      },
      "TaskFrequencyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "daysOfWeek": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "dayOfMonth": {
            "type": "object"
          },
          "interval": {
            "type": "number"
          },
          "timesPerPeriod": {
            "type": "object"
          },
          "endDate": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "type",
          "daysOfWeek",
          "interval"
        ]
      },
      "PomodoroSettingsResponseDto": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string"
          },
          "focusDuration": {
            "type": "number"
          },
          "breakDuration": {
            "type": "number"
          },
          "longBreakDuration": {
            "type": "number"
          },
          "totalSessions": {
            "type": "number"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "autoStartBreaks": {
            "type": "boolean"
          },
          "autoStartFocus": {
            "type": "boolean"
          }
        },
        "required": [
          "taskId",
          "focusDuration",
          "breakDuration",
          "longBreakDuration",
          "totalSessions",
          "isEnabled",
          "autoStartBreaks",
          "autoStartFocus"
        ]
      },
      "TodoSubtaskResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "isCompleted": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "title",
          "isCompleted"
        ]
      },
      "TaskResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "categoryId": {
            "type": "object"
          },
          "challengeId": {
            "type": "object"
          },
          "frequencyId": {
            "type": "object"
          },
          "title": {
            "type": "string"
          },
          "taskType": {
            "type": "string",
            "enum": [
              "HABIT",
              "ROUTINE",
              "TODO",
              "MINDSET"
            ]
          },
          "iconName": {
            "type": "object"
          },
          "colorValue": {
            "type": "object"
          },
          "imageUrl": {
            "type": "object"
          },
          "isPublic": {
            "type": "boolean"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "isPredefined": {
            "type": "boolean"
          },
          "forkCount": {
            "type": "number"
          },
          "isForked": {
            "type": "boolean"
          },
          "originalTaskId": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "type": "object"
          },
          "startDate": {
            "format": "date-time",
            "type": "string"
          },
          "endDate": {
            "type": "object"
          },
          "notes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reminders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskReminderResponseDto"
            }
          },
          "frequency": {
            "$ref": "#/components/schemas/TaskFrequencyResponseDto"
          },
          "pomodoroSettings": {
            "$ref": "#/components/schemas/PomodoroSettingsResponseDto"
          },
          "goalValue": {
            "type": "number"
          },
          "currentValue": {
            "type": "number"
          },
          "unitId": {
            "type": "object"
          },
          "direction": {
            "type": "string",
            "enum": [
              "POSITIVE",
              "NEGATIVE"
            ]
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "startTime": {
            "type": "object"
          },
          "dueTime": {
            "type": "object"
          },
          "priority": {
            "type": "string",
            "enum": [
              "NONE",
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ]
          },
          "isFlagged": {
            "type": "boolean"
          },
          "url": {
            "type": "object"
          },
          "subtasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TodoSubtaskResponseDto"
            }
          },
          "affirmation": {
            "type": "object"
          },
          "durationMinutes": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "userId",
          "title",
          "taskType",
          "isPublic",
          "isDeleted",
          "isPredefined",
          "forkCount",
          "isForked",
          "createdAt",
          "updatedAt",
          "startDate",
          "notes",
          "reminders"
        ]
      },
      "PaginatedTasksResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskResponseDto"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ]
      },
      "LogCompletionDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "description": "Numeric value (e.g. for habits)"
          },
          "status": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "TaskCompletionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "taskId": {
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string"
          },
          "value": {
            "type": "object"
          },
          "status": {
            "type": "string"
          },
          "notes": {
            "type": "object"
          },
          "description": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "taskId",
          "completedAt",
          "status"
        ]
      },
      "PaginatedCompletionsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskCompletionResponseDto"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "categoryId": {
            "type": "object"
          },
          "iconName": {
            "type": "object"
          },
          "colorValue": {
            "type": "object"
          },
          "imageUrl": {
            "type": "object"
          },
          "isPublic": {
            "type": "boolean"
          },
          "startDate": {
            "format": "date-time",
            "type": "string"
          },
          "endDate": {
            "type": "object"
          },
          "notes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "goalValue": {
            "type": "number"
          },
          "currentValue": {
            "type": "number"
          },
          "unitId": {
            "type": "object"
          },
          "direction": {
            "type": "string"
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "startTime": {
            "type": "object"
          },
          "dueTime": {
            "type": "object"
          },
          "priority": {
            "type": "string"
          },
          "isFlagged": {
            "type": "boolean"
          },
          "url": {
            "type": "object"
          },
          "affirmation": {
            "type": "object"
          },
          "durationMinutes": {
            "type": "object"
          },
          "reminders": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subtasks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "frequency": {
            "type": "object"
          },
          "pomodoroSettings": {
            "type": "object"
          }
        }
      },
...
    }
  }
}